Start SappyIntegration
